#1, #5 add Limit and OrderBook types, add Dao for Trader, rename tests

+++ client/model/User.cpp
@@ -1,6 +1,6 @@
 #include "User.hpp"
 
-User::User(unsigned long id, std::string username)
+User::User(size_t id, std::string username)
     : id(id),
       username(username) {}
